chore: refactor to allow build on linux and non-linux hosts
